Changed damage values to floats
- works better with stat system

+++ Combat/DamageInfo.cs
@@ -5,13 +5,13 @@ namespace PolyWare.Combat {
 	[Serializable]
 	public struct DamageInfo {
 		[HideInInspector] public readonly bool FromPlayer;
-		public int Damage;
+		public float Damage;
 		public Element Element;
 		public bool PenetrateShields;
 
 		public DamageInfo(bool isPlayer, DamageInfo damageInfo) : this(isPlayer, damageInfo.Element, damageInfo.Damage, damageInfo.PenetrateShields) { }
 
-		public DamageInfo(bool fromPlayer, Element element, int damage, bool penetrateShields) {
+		public DamageInfo(bool fromPlayer, Element element, float damage, bool penetrateShields) {
 			FromPlayer = fromPlayer;
 			Element = element;
 			Damage = damage;
